search indexes moved to gnuma.models
