+++ backend/gnuma/models.py
@@ -1,5 +1,13 @@
+# django imports
 from django.db import models
 from django.contrib.auth.models import User
+
+# OS imports
+import datetime
+
+# Haystack
+from haystack import indexes
+
 
 
 
@@ -136,3 +144,33 @@ class Report(models.Model):
     def __str__(self):
         return self.sender.__str__() + " " + self.content
 
+#
+# search indexes
+#
+# python imports
+class BookIndex(indexes.SearchIndex, indexes.Indexable):
+    text = indexes.CharField(document = True, use_template = True)
+    author = indexes.CharField(model_attr = 'author')
+    #
+    # renderer
+    # 
+
+    title = indexes.EdgeNgramField(model_attr = 'title')
+    isbn = indexes.EdgeNgramField(model_attr = 'isbn')
+    
+    def get_model(self):
+        return Book
+
+    def index_queryset(self, using = None):
+        return self.get_model().objects.all()
+
+
+class OfficeIndex(indexes.SearchIndex, indexes.Indexable):
+    text = indexes.EdgeNgramField(document = True, model_attr = 'name')
+
+
+    def get_model(self):
+        return Office
+
+    def index_queryset(self, using = None):
+        return self.get_model().objects.all()
